agente contenido y rrss/requirements.txt
# =============================================================================
# AGENTE CONTENIDO Y RRSS - Dependencias Python
# =============================================================================
# Instalar: pip install -r requirements.txt
# =============================================================================

# --- Google Trends ---
trendspyg>=0.1.0

# --- Web Scraping ---
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
httpx>=0.27.0

# --- Procesamiento de Datos ---
pandas>=2.2.0
rapidfuzz>=3.6.0
unidecode>=1.3.0

# --- Configuracion y Entorno ---
pyyaml>=6.0.0
python-dotenv>=1.0.0
pydantic>=2.6.0

# --- Google Drive ---
PyDrive2>=1.19.0
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0

# --- Generacion de Imagenes ---
Pillow>=10.2.0

# --- Scheduling ---
APScheduler>=3.10.0

# --- Logging ---
colorlog>=6.8.0

# --- Testing ---
pytest>=8.0.0
pytest-cov>=4.1.0
responses>=0.25.0
